tests: the by-move tail gate must take a struct-literal factory
The counter-factual is the leak: recur-local-mutual-factory measures the
everyday letrec-of-closures factory, whose letrec body's tail is a struct
literal over both members. It reads open today, because the by-move gate
refuses a member carried into any non-member tail call.

merge_admits_member_passed_by_move_to_native_tail pins the merge itself, and
test_native_constant_binding_never_replaces_the_frame pins the predicate the
gate asks: a native constant keeps the frame, a mutated binding does not.

The guardfree fixture gains both by-move rows — the factory, re-entered through
the member the struct handed out after the builder's frame is gone, and its
counterweight, a member handed to a callee that can replace the frame.

The merge test file held two subjects and had reached the reading budget, so
the tail-callee gates move to their own tailgate module.
